requirements.txt
streamlit
google-generativeai
python-dotenv
IPython
pathlib
textwrap
PIL